backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.17
python-magic==0.4.27
Pillow==11.0.0
minio==7.2.10
python-jose[cryptography]==3.3.0
bcrypt
httpx==0.27.2
email-validator==2.2.0
celery==5.4.0
redis==5.0.1
pytest==8.3.3
mediapipe==0.10.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
google-genai
